requirements.txt
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.2
matplotlib==3.7.3
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.1
scikit-learn==1.3.1